Move Field tests to individual component suites
